refactor: create instagib include proxy files
