+++ CMakeLists.txt
@@ -2899,6 +2899,14 @@ if(SERVER)
     instagib/gamecontext/bangcommands.cpp
     instagib/gamecontroller.cpp
     instagib/gamecontroller.h
+    instagib/includes/character.h
+    instagib/includes/chat_commands.h
+    instagib/includes/config_variables.h
+    instagib/includes/gamecontext.h
+    instagib/includes/gamecontroller.h
+    instagib/includes/persistent_client_data.h
+    instagib/includes/player.h
+    instagib/includes/rcon_commands.h
     instagib/ip_storage.cpp
     instagib/ip_storage.h
     instagib/laser_text.cpp
