+++ src/engine/shared/config_variables.h
@@ -784,4 +784,4 @@ MACRO_CONFIG_INT(ClVideoRecorderFPS, cl_video_recorder_fps, 60, 1, 1000, CFGFLAG
  */
 
 // ddnet-insta
-#include "config_variables_insta.h"
+#include <game/server/instagib/includes/config_variables.h>
