+++ src/game/server/instagib/includes/chat_commands.h
@@ -1,0 +1,6 @@
+// This file can be included several times.
+
+#include <game/server/instagib/chat_commands.h>
+
+// Put your custom includes below this comment
+// to avoid merge conflicts when this file gets updated.
