+++ src/game/server/instagib/includes/config_variables.h
@@ -1,0 +1,6 @@
+// This file can be included several times.
+
+#include <engine/shared/config_variables_insta.h>
+
+// Put your custom includes below this comment
+// to avoid merge conflicts when this file gets updated.
