+++ src/game/server/instagib/includes/persistent_client_data.h
@@ -1,0 +1,6 @@
+// This file can be included several times.
+
+#include <game/server/instagib/persistent_client_data.h>
+
+// Put your custom includes below this comment
+// to avoid merge conflicts when this file gets updated.
